Added a conditions file
Conditions file will be ran at the end of every loop. It checks to see if something should actually happen in the game. For example, if you need to collect a mug, a book, and a chalice to open a door, this file will check your inventory for that and if the conditions are met, that door will open. Needs comments.